Change registeset_type_marker underlying type to uint8_t

+++ include/viua/bytecode/bytetypedef.h
@@ -42,7 +42,7 @@ namespace viua {
 
             typedef uint32_t timeout;
 
-            typedef uint32_t registerset_type_marker;
+            typedef uint8_t registerset_type_marker;
         }
     }
 }
